Move ahdb to helper method
